adoptd to mangling of internal functions in scala3
