+++ jsoniter-scala-macros/jvm/src/test/scala-3/JsonCodeMakerJVMSpec.scala
@@ -5,6 +5,10 @@ import com.github.plokhotnyuk.jsoniter_scala.macros.JsonCodecMaker._
 
 import scala.annotation.tailrec
 
+
+//TODO:  java-only enums
+
+
 class JsonCodecMakerJVMSpec extends VerifyingSpec {
 
   "JsonCodecMaker.make generate codes which" should {
@@ -18,14 +22,14 @@ class JsonCodecMakerJVMSpec extends VerifyingSpec {
 
       implicit val codecOfNestedStructs: JsonValueCodec[Nested] = make(CodecMakerConfig.withAllowRecursiveTypes(true))
       val bytes = ("{" + "\"n\":{" * 1000000 + "}" * 1000000 + "}").getBytes
-      val readStackTrace = AssertionUtils.assertStackOverflow(readFromArray[Nested](bytes))
-      assert(readStackTrace.contains(".d0("))
-      assert(!readStackTrace.contains(".d1("))
-      assert(!readStackTrace.contains(".decodeValue("))
-      val writeStackTrace = AssertionUtils.assertStackOverflow(writeToArray[Nested](construct()))
-      assert(writeStackTrace.contains(".e0("))
-      assert(!writeStackTrace.contains(".e1("))
-      assert(!writeStackTrace.contains(".encodeValue("))
+      val readStackTrace = AssertionUtils.assertStackOverflow(readFromArray[Nested](bytes)).split('\n')
+      assert(readStackTrace.exists(_.matches(".*\\.d0(\\$[0-9]+)*\\(.*")))
+      assert(!readStackTrace.exists(_.matches(".*\\.d1(\\$[0-9]+)*\\(.*")))
+      assert(!readStackTrace.exists(_.matches(".*\\.decodeValue(\\$[0-9]+)*\\(")))
+      val writeStackTrace = AssertionUtils.assertStackOverflow(writeToArray[Nested](construct())).split('\n')
+      assert(writeStackTrace.exists(_.matches(".*\\.e0(\\$[0-9]+)*\\(.*")))
+      assert(!writeStackTrace.exists(_.matches(".*\\.e1(\\$[0-9]+)*\\(.*")))
+      assert(!writeStackTrace.exists(_.matches(".*\\.encodeValue(\\$[0-9]+)*\\(.*")))
     }
   }
 }
